+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [2.3.1] - 2018-03-28
+
+### Added
+- LocationUtil::computeCoordinatesByArray()
+- LocationUtil::computeCoordinatesByString()
+
 ## [2.3.0] - 2018-03-28
 
 ### Added
 - ArrayUtil::removePrefix()
 - DatabsseCacheUtil
 - StringUtil::removeLeadingString()
+- LocationUtil
 
 ## [2.2.2] - 2018-03-27